Backend_assingment/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
apify_client==1.9.2
apify_shared==1.3.1
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
Django==4.2.20
django-appconf==1.1.0
django-cryptography==1.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.9.0
more-itertools==10.6.0
oauthlib==3.2.2
openai==1.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
